+++ CHANGELOG.md
@@ -8,7 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Fixed
+
 - Fix admin dialog: refresh Project Schema list after a connection error.
+- Fix admin dialog: list all WS/UD schemas on the active connection.
 
 ## [4.15.3] - 2026-07-17