+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [4.15.4] - 2026-07-21
+
 ### Fixed
 
 - Fix admin dialog: refresh Project Schema list after a connection error.
@@ -525,7 +527,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Form change detection and caching improvements.
 - Large-scale flake8 and typing standardization.
 
-[unreleased]: https://github.com/giswater/plugin/compare/v4.15.3...release/4.15
+[unreleased]: https://github.com/giswater/plugin/compare/v4.15.4...release/4.15
+[4.15.4]: https://github.com/giswater/plugin/compare/v4.15.3...v4.15.4
 [4.15.3]: https://github.com/giswater/plugin/compare/v4.15.2...v4.15.3
 [4.15.2]: https://github.com/giswater/plugin/compare/v4.15.1...v4.15.2
 [4.15.1]: https://github.com/giswater/plugin/compare/v4.15.0...v4.15.1